walletconnect dailog -> page
