+++ src/config/routes.ts
@@ -31,4 +31,7 @@ export const AppRoutes = {
   welcome: {
     index: '/welcome',
   },
+  walletConnect: {
+    transaction: '/wallet-connect/transaction',
+  },
 }
